Vickrey-Dutch Auction Crowdsale
AuctionSale is a base contract for managing a token sale based on the result of a previous auction. Investors can make token purchases based on the result of an auction and AuctionSale will assign them tokens based on said previous auction. 
Funds collected are forwarded to a wallet as they arrive.